report success for the watch.sqlquery.{query} service
This is the service we also report errors for. THe problem is, that we only did
report errors and never an 'ok' state.

Riemann consequently warned about the service being, but from riemann's point of
view it never came back up again.

I wasn't too happy with the solution back then for reporting errors and I'm
still not happy with it. Maybe we should pass an explicit parameter as service
name that gets reported.

fixes https://github.com/trustlines-network/watch/issues/6